Plot3d option added + directlabels from imports to depends + corrected error in package documentation file
Added a plot3d option for plotting mf2darray objects in 3D with rgl. The directlabels package was moved from imports to depends because of the function from the directlabels depends that was missing by just importing. The text in the package documentation file that was not in a section was now moved in the corresponding section to solve the errors when building the package.